# apps/platform-api/specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog API Documentation",
    "version": "1.0.0",
    "description": "API Documentation"
  },
  "paths": {
...
    "/api/v1/permission": {
      "get": {
        "summary": "列出所有权限",
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "权限列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "description": {
                            "type": "object"
                          },
                          "created_at": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "code",
                          "description",
                          "created_at"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建或更新权限",
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "code": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "权限创建/更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "affectedRows": {
                      "type": "number"
                    },
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "affectedRows"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除权限",
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_id": {
                    "type": "number",
                    "minimum": -9007199254740991,
                    "maximum": 9007199254740991
                  },
                  "target_user_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "permission_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "权限删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "affectedRows": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "affectedRows"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission-request": {
      "post": {
        "summary": "创建权限请求",
        "tags": [
          "PermissionRequests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_ids": {
                    "type": "array",
                    "items": {
                      "type": "number",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "permission_ids"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "权限请求创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "user_id": {
                      "type": "string"
                    },
                    "permission_ids": {
                      "type": "array",
                      "items": {
                        "type": "number"
                      }
                    },
                    "request_type": {
                      "type": "string",
                      "enum": [
                        "PERMISSION",
                        "ACCESS"
                      ]
                    },
                    "target_user_id": {
                      "type": "object"
                    },
                    "path": {
                      "type": "object"
                    },
                    "scope": {
                      "type": "object"
                    },
                    "entity_type": {
                      "type": "object"
                    },
                    "entity_data": {
                      "type": "object"
                    },
                    "expires_at": {
                      "type": "object"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "APPROVED",
                        "REJECTED"
                      ]
                    },
                    "reason": {
                      "type": "object"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "user_id",
                    "permission_ids",
                    "request_type",
                    "target_user_id",
                    "path",
                    "scope",
                    "entity_type",
                    "entity_data",
                    "expires_at",
                    "status",
                    "reason",
                    "created_at",
                    "updated_at"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "列出所有权限请求",
        "tags": [
          "PermissionRequests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "权限请求列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "user_id": {
                            "type": "string"
                          },
                          "permission_ids": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            }
                          },
                          "request_type": {
                            "type": "string",
                            "enum": [
                              "PERMISSION",
                              "ACCESS"
                            ]
                          },
                          "target_user_id": {
                            "type": "object"
                          },
                          "path": {
                            "type": "object"
                          },
                          "scope": {
                            "type": "object"
                          },
                          "entity_type": {
                            "type": "object"
                          },
                          "entity_data": {
                            "type": "object"
                          },
                          "expires_at": {
                            "type": "object"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "PENDING",
                              "APPROVED",
                              "REJECTED"
                            ]
                          },
                          "reason": {
                            "type": "object"
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "updated_at": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "user_id",
                          "permission_ids",
                          "request_type",
                          "target_user_id",
                          "path",
                          "scope",
                          "entity_type",
                          "entity_data",
                          "expires_at",
                          "status",
                          "reason",
                          "created_at",
                          "updated_at"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission-request/me": {
      "get": {
        "summary": "列出我的权限请求",
        "tags": [
          "PermissionRequests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "我的权限请求列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "user_id": {
                        "type": "string"
                      },
                      "permission_ids": {
                        "type": "array",
                        "items": {
                          "type": "number"
                        }
                      },
                      "request_type": {
                        "type": "string",
                        "enum": [
                          "PERMISSION",
                          "ACCESS"
                        ]
                      },
                      "target_user_id": {
                        "type": "object"
                      },
                      "path": {
                        "type": "object"
                      },
                      "scope": {
                        "type": "object"
                      },
                      "entity_type": {
                        "type": "object"
                      },
                      "entity_data": {
                        "type": "object"
                      },
                      "expires_at": {
                        "type": "object"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "PENDING",
                          "APPROVED",
                          "REJECTED"
                        ]
                      },
                      "reason": {
                        "type": "object"
                      },
                      "created_at": {
                        "type": "string"
                      },
                      "updated_at": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "user_id",
                      "permission_ids",
                      "request_type",
                      "target_user_id",
                      "path",
                      "scope",
                      "entity_type",
                      "entity_data",
                      "expires_at",
                      "status",
                      "reason",
                      "created_at",
                      "updated_at"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission-request/{id}/approve": {
      "post": {
        "summary": "批准权限请求",
        "tags": [
          "PermissionRequests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "权限请求已批准",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "user_id": {
                      "type": "string"
                    },
                    "permission_ids": {
                      "type": "array",
                      "items": {
                        "type": "number"
                      }
                    },
                    "request_type": {
                      "type": "string",
                      "enum": [
                        "PERMISSION",
                        "ACCESS"
                      ]
                    },
                    "target_user_id": {
                      "type": "object"
                    },
                    "path": {
                      "type": "object"
                    },
                    "scope": {
                      "type": "object"
                    },
                    "entity_type": {
                      "type": "object"
                    },
                    "entity_data": {
                      "type": "object"
                    },
                    "expires_at": {
                      "type": "object"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "APPROVED",
                        "REJECTED"
                      ]
                    },
                    "reason": {
                      "type": "object"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "user_id",
                    "permission_ids",
                    "request_type",
                    "target_user_id",
                    "path",
                    "scope",
                    "entity_type",
                    "entity_data",
                    "expires_at",
                    "status",
                    "reason",
                    "created_at",
                    "updated_at"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission-request/{id}/reject": {
      "post": {
        "summary": "拒绝权限请求",
        "tags": [
          "PermissionRequests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "权限请求已拒绝",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "user_id": {
                      "type": "string"
                    },
                    "permission_ids": {
                      "type": "array",
                      "items": {
                        "type": "number"
                      }
                    },
                    "request_type": {
                      "type": "string",
                      "enum": [
                        "PERMISSION",
                        "ACCESS"
                      ]
                    },
                    "target_user_id": {
                      "type": "object"
                    },
                    "path": {
                      "type": "object"
                    },
                    "scope": {
                      "type": "object"
                    },
                    "entity_type": {
                      "type": "object"
                    },
                    "entity_data": {
                      "type": "object"
                    },
                    "expires_at": {
                      "type": "object"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "APPROVED",
                        "REJECTED"
                      ]
                    },
                    "reason": {
                      "type": "object"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "user_id",
                    "permission_ids",
                    "request_type",
                    "target_user_id",
                    "path",
                    "scope",
                    "entity_type",
                    "entity_data",
                    "expires_at",
                    "status",
                    "reason",
                    "created_at",
                    "updated_at"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission-request/{id}": {
      "delete": {
        "summary": "删除权限请求",
        "tags": [
          "PermissionRequests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "权限请求删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "affectedRows": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "affectedRows"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}